Added a check to fix 'Cannot override' bug
This one should fix issue #814 once and for all. The issue was caused by the SSDT that was adding _DSM Method to a Package (forbidden by ACPI Specification) defined on the DSDT (\_SB.PR00). The new if-condition checks if the object is a Processor. By doing this way, if it finds a Package it doesn't add the _DSM Method.