Add RTC0 patch for ASUS X299 System
Asus forgot to map the 0x72 and 0x73 regions in RTC device. Look in my original DSDT and find RTC. 
Then see what regions the CRS is missing, for me you can see the length is 0x02 meaning it only contains 0x70 and 0x71.
and the next region starts at 0x74 meaning we have 2 regions missing

so, MykolaG was find this value 0x04.