Reset timeout on cursor interaction, spelling fix
My keyboard is connected to a Bluetooth reciever and therefore doesn't work.
Sadllu, mu Logitech G603 + 2.4Ghz reciever is a tad laggy and this results in me 50% of the time not being able to select what I'd like to boot up. 

So I added the feature of also disabling the timeout on pointer events. 

Does this make sense? Please review this carefully. Was a looong time ago I wrote any C code.. :)

Also fixed a spelling error.