OpenCoreUefi: Remove DeduplicateBootOrder quirk
This quirk was required earlier to workaround ASUS bios issues.
A better workaround for these issues was added in 0.6.4.
At this point DeduplicateBootOrder may only mask issues related
to incorrect Bootstrap/BootstrapShort preference for the board
and is thus removed due to being essentially harmful.

Note, that for some firmwares it is normal to have 2 Windows
entries, and DeduplicateBootOrder had nothing to do with it.