OpenCanopy: Do not register 'Click' events with AppleEvent
We do not need the concept of a 'Click', which is a short-term button down-up sequence in a very narrow square area, as it is common behaviour to invoke any button on button up if and only if the cursor is in the button area. Timing and relative movement do not matter for the action.