OpenDuet: Resolve issue starting OS since new secure PE/COFF loader integration
Prior to this change, but since PE/COFF loader integration, RuntimeDxe
(which includes runtime relocation support) was built with UE support
only (because the previous - now revised - default for non-FV in MdePkg
was UE only, and this file overrides the default for FV to UE only)
however DxeCore was built to load PE and UE, so PE drivers with runtime
support (e.g. OpenRuntime.efi) would load but then ASSERT at runtime
relocation time (if asserts were enabled; otherwise failing later, on
OS callback to the runtime).

In this revised version (also taking into account the matched update to
AUDK, which changes FV and non-FV defaults to be PE-only) OpenDuetPkg
whole-platform support is now set to UE-only in FV and PE-only in non-FV
(therefore RuntimeDxe is aware that both formats may be seen at relocation
time).

PcdUefiImageFormatSupport[Non]Fv values should be (and in this update
are) defined globally for the entire platform, not per-module.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>
Signed-off-by: Marvin Häuser <mhaeuser@posteo.de>